data-push connection manager refactor (part 2): reorg
