+++ legend-engine-xts-authentication/legend-engine-xt-authentication-experimental/pom.xml
@@ -19,7 +19,7 @@
     <parent>
         <groupId>org.finos.legend.engine</groupId>
         <artifactId>legend-engine-xts-authentication</artifactId>
-        <version>4.25.1-SNAPSHOT</version>
+        <version>4.23.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -139,7 +139,3 @@
         </dependency>
 
     </dependencies>
-
-
-
-</project>
