--- legend-engine-xts-authentication/legend-engine-xt-authentication-experimental/src/main/resources/META-INF/services/org.finos.legend.connection.jdbc.JDBCConnectionManager
@@ -1,4 +1,0 @@
-org.finos.legend.connection.jdbc.drivers.H2_JDBCConnectionManager
-org.finos.legend.connection.jdbc.drivers.PostgreSQL_JDBCConnectionManager
-org.finos.legend.connection.jdbc.drivers.MemSQL_JDBCConnectionManager
-org.finos.legend.connection.jdbc.drivers.SQLServer_JDBCConnectionManager
