+++ legend-engine-xts-authentication/legend-engine-xt-authentication-implementation-core/src/main/java/org/finos/legend/authentication/intermediationrule/impl/UserPasswordFromVaultRule.java
@@ -34,5 +34,4 @@ public class UserPasswordFromVaultRule extends IntermediationRule<UserPasswordAu
         String password = super.lookupSecret(specification.password);
         return new PlaintextUserPasswordCredential(specification.username, password);
     }
-
 }
