+++ legend-engine-xts-data-push/pom.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>legend-engine</artifactId>
         <groupId>org.finos.legend.engine</groupId>
@@ -25,12 +26,13 @@
                 <groupId>com.github.ferstl</groupId>
                 <artifactId>depgraph-maven-plugin</artifactId>
                 <configuration>
-                    <includes>*:legend-engine-xt-data*,*:legend-engine-store*,*:legend-engine-server-support*,*:legend-engine-xt-relationalStore*,*:legend-engine-xt-authentication*</includes>
+                    <includes>
+                        *:legend-engine-xt-data*,*:legend-engine-store*,*:legend-engine-server-support*,*:legend-engine-xt-relationalStore*,*:legend-engine-xt-authentication*
+                    </includes>
                     <excludes>*:*:test</excludes>
                     <graphFormat>text</graphFormat>
                 </configuration>
             </plugin>
         </plugins>
     </build>
-
 </project>
