+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-connection/src/main/resources/META-INF/services/org.finos.legend.connection.jdbc.JDBCConnectionManager
@@ -1,0 +1,1 @@
+org.finos.legend.connection.jdbc.driver.H2_JDBCConnectionManager
