+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-memsql/legend-engine-xt-relationalStore-memsql-connection/src/main/resources/META-INF/services/org.finos.legend.connection.jdbc.JDBCConnectionManager
@@ -1,0 +1,1 @@
+org.finos.legend.connection.jdbc.driver.MemSQL_JDBCConnectionManager
