+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-memsql/pom.xml
@@ -28,6 +28,7 @@
 
     <modules>
         <module>legend-engine-xt-relationalStore-memsql-pure</module>
+        <module>legend-engine-xt-relationalStore-memsql-connection</module>
         <module>legend-engine-xt-relationalStore-memsql-execution</module>
         <module>legend-engine-xt-relationalStore-memsql-execution-tests</module>
     </modules>
