+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-sqlserver/legend-engine-xt-relationalStore-sqlserver-connection/src/main/resources/META-INF/services/org.finos.legend.connection.jdbc.JDBCConnectionManager
@@ -1,0 +1,1 @@
+org.finos.legend.connection.jdbc.driver.SQLServer_JDBCConnectionManager
