+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-sqlserver/pom.xml
@@ -27,6 +27,7 @@
     <name>Legend Engine - XT - Relational Store - DB Extension - SQL Server</name>
 
     <modules>
+        <module>legend-engine-xt-relationalStore-sqlserver-connection</module>
         <module>legend-engine-xt-relationalStore-sqlserver-execution</module>
         <module>legend-engine-xt-relationalStore-sqlserver-execution-tests</module>
         <module>legend-engine-xt-relationalStore-sqlserver-pure</module>
